Revu l'initialisation de verb_gc.  Ajout de messages de trace pour attempt_open
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@3419 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/misc.h
@@ -56,7 +56,7 @@ void fatal_error_arg (char *, char *) Noreturn;
 
 /* GC flags and messages */
 
-extern int verb_gc;
+extern unsigned long verb_gc;
 void gc_message (int, char *, unsigned long);
 
 /* Memory routines */
